Move WriteWatchersWithCap to the Monitor namespace.
