+++ src/modules/m_ircv3.cpp
@@ -153,7 +153,7 @@ class ModuleIRCv3
 		msg.PushParamRef(param);
 		ClientProtocol::Event accountevent(accountprotoev, msg);
 		IRCv3::WriteNeighborsWithCap res(user, accountevent, cap_accountnotify, true);
-		IRCv3::WriteWatchersWithCap(monitorapi, user, accountevent, cap_accountnotify, res.GetAlreadySentId());
+		Monitor::WriteWatchersWithCap(monitorapi, user, accountevent, cap_accountnotify, res.GetAlreadySentId());
 	}
 
 	void OnUserAway(User* user) CXX11_OVERRIDE
@@ -165,7 +165,7 @@ class ModuleIRCv3
 		AwayMessage msg(user);
 		ClientProtocol::Event awayevent(joinhook.awayprotoev, msg);
 		IRCv3::WriteNeighborsWithCap res(user, awayevent, joinhook.awaycap);
-		IRCv3::WriteWatchersWithCap(monitorapi, user, awayevent, joinhook.awaycap, res.GetAlreadySentId());
+		Monitor::WriteWatchersWithCap(monitorapi, user, awayevent, joinhook.awaycap, res.GetAlreadySentId());
 	}
 
 	void OnUserBack(User* user) CXX11_OVERRIDE
